# public/openapi.yaml
openapi: 3.0.0
info:
  title: Visamonk University Chatbot API
  version: 1.0.0
  description: API documentation for the Visamonk University Chatbot
servers:
  - url: https://cuddly-space-couscous-x59xq6pg5v4qhvqq7-3000.app.github.dev/api
    description: Codespaces development server
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Unauthorized
  /auth/verify:
    get:
      summary: Verify JWT token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  success:
                    type: boolean
        '401':
          description: Invalid token
...